Updated admin files to use new sql procedure

+++ Spoofy/admin/adminship.php
@@ -1,6 +1,7 @@
 <!-- @todo: We can turn this into an actual button at some point instead of a hyperlink -->
 <?php
 include "../modules/mysql_connect.php";
+require_once("../modules/python_connect.php");
 
 if(!isset($_SESSION)) { session_start(); }
 if (isset($_SESSION["LoggedIn"]) && $_SESSION["LoggedIn"] && $_SESSION["Admin"]) {
@@ -15,23 +16,31 @@ if (isset($_SESSION["LoggedIn"]) && $_SESSION["LoggedIn"] && $_SESSION["Admin"])
             $prepare -> execute();
             $result = $prepare -> get_result();
             if (mysqli_num_rows($result) == 0) {
-                $sql = "INSERT INTO ADMIN (AdminID) VALUES (?)";
-                $prepare = mysqli_prepare($con, $sql);
-                if ($prepare) {
-                    $prepare -> bind_param("s", $UserID);
-                    $prepare -> execute();
-                }
+                $sql = "INSERT INTO ADMIN (AdminID) VALUES ($UserID)";
+                sendQuery($sql);
+                // $sql = "INSERT INTO ADMIN (AdminID) VALUES (?)";
+                // $prepare = mysqli_prepare($con, $sql);
+                // if ($prepare) {
+                //     $prepare -> bind_param("s", $UserID);
+                //     $prepare -> execute();
+                // }
             }
         }
-        $prepare -> close();
-    } else {
-        $sql = "DELETE FROM ADMIN WHERE AdminID=?";
-        $prepare = mysqli_prepare($con, $sql);
         if ($prepare) {
-            $prepare -> bind_param("s", $UserID);
-            $prepare -> execute();
+            $prepare -> close();
         }
-        $prepare -> close();
+    } else {
+        $sql = "DELETE FROM ADMIN WHERE AdminID=$UserID";
+        sendQuery($sql);
+        // $sql = "DELETE FROM ADMIN WHERE AdminID=?";
+        // $prepare = mysqli_prepare($con, $sql);
+        // if ($prepare) {
+        //     $prepare -> bind_param("s", $UserID);
+        //     $prepare -> execute();
+        // }
+        if ($prepare) {
+            $prepare -> close();
+        }
     }
     header("location: /admin/manage_users.php");
 } else {
